fixed gcc warning(s)
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@21367 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/dsql/pass1.cpp
@@ -2351,8 +2351,8 @@ static void field_duplication(const TEXT* qualifier_name, const TEXT* field_name
 
 	if (qualifier_name)
 	{
-		sprintf(field_buffer, "%.*s.%.*s", MAX_SQL_IDENTIFIER_LEN, qualifier_name,
-				MAX_SQL_IDENTIFIER_LEN, field_name);
+		sprintf(field_buffer, "%.*s.%.*s", (int) MAX_SQL_IDENTIFIER_LEN, qualifier_name,
+				(int) MAX_SQL_IDENTIFIER_LEN, field_name);
 		field_name = field_buffer;
 	}
 
@@ -2386,8 +2386,8 @@ static void field_unknown(const TEXT* qualifier_name, const TEXT* field_name,
 
 	if (qualifier_name)
 	{
-		sprintf(field_buffer, "%.*s.%.*s", MAX_SQL_IDENTIFIER_LEN, qualifier_name,
-				MAX_SQL_IDENTIFIER_LEN, field_name ? field_name : "*");
+		sprintf(field_buffer, "%.*s.%.*s", (int) MAX_SQL_IDENTIFIER_LEN, qualifier_name,
+				(int) MAX_SQL_IDENTIFIER_LEN, field_name ? field_name : "*");
 		field_name = field_buffer;
 	}
